+++ packages/language-server/src/language-service/services/completion/constants.ts
@@ -1,4 +1,5 @@
 export const RUNTIMES = [
+	"nodejs",
 	"nodejs10.x",
 	"nodejs8.10",
 	"nodejs6.10",