+++ packages/vscode/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## v0.5.0
+- Various schemas bugfixes and improvements for Cloudformation and Serverless framework
+- Autocompletion for conditions properties and functions
+- Minor bug fixes
+
 ## v0.4.12
 - Allow OPTION as a valid HTTP method in Serverless framework schema